requirements.txt
# Core
torch>=1.7.0
opencv-python>=4.5.1.48
numpy>=1.19.5
pygame>=2.1.0
pillow>=8.1.0
matplotlib>=3.3.4

# Optional but common for YOLOv5
scipy>=1.4.1
tqdm>=4.64.0
requests>=2.23.0
pyyaml>=5.3.1